Cloud_Functions_PubSub/process-payment-failure/requirements.txt
google-cloud-pubsub
google-cloud-firestore
google-cloud-logging==3.7.0